PlayStation 4 support, Vibration fixes, improvements for tr_, reduce memory leaks.

+++ SonicTimeTwisted.gmx/scripts/shift_savefiles.gml
@@ -1,21 +1,27 @@
 // if a save 0 exists, it means that the game has been previously played using a 
 // version that didn't have the "No Save" slot yet.
 // In this case, shift all save files to the right.
-var zeroFilename = save_file_name(0);
-if(file_exists(zeroFilename))
+if (DEVICE_INFO & DEVICE_OS_PLAYSTATION) return false;
+else
 {
-    // if a save file somehow exists at the very end, then it will be deleted, unfortunately
-    var lastFilename = save_file_name(7);
-    if(file_exists(lastFilename))
+    var zeroFilename = save_file_name(0);
+    if(file_exists(zeroFilename))
     {
-        file_delete(lastFilename);
-    }
-    // we are shifting everything!
-    for(var i = 6; i >= 0; i--)
-    {
-        if(file_exists(save_file_name(i)))
+        // if a save file somehow exists at the very end, then it will be deleted, unfortunately
+        var lastFilename = save_file_name(7);
+        if(file_exists(lastFilename))
         {
-            file_rename(save_file_name(i), save_file_name(i+1));
+            file_delete(lastFilename);
+        }
+        // we are shifting everything!
+        for(var i = 6; i >= 0; i--)
+        {
+            if(file_exists(save_file_name(i)))
+            {
+                file_rename(save_file_name(i), save_file_name(i+1));
+            }
         }
     }
+    
+    return true;
 }
